+++ src/test/test_rate_limiter.ts
@@ -136,7 +136,8 @@ describe('Rate Limiter', function() {
       const afterJobs = after(numJobs, () => {
         try {
           const timeDiff = Date.now() - startTime;
-          expect(timeDiff).to.be.gte(numGroups * 1000);
+          // In some test envs, these timestamps can drift.
+          expect(timeDiff).to.be.gte(numGroups * 990);
           expect(timeDiff).to.be.below((numGroups + 1) * 1100);
 
           for (const group in completed) {
@@ -144,7 +145,7 @@ describe('Rate Limiter', function() {
             for (let i = 1; i < completed[group].length; i++) {
               const diff = completed[group][i] - prevTime;
               expect(diff).to.be.below(2000);
-              expect(diff).to.be.gte(1000);
+              expect(diff).to.be.gte(990);
               prevTime = completed[group][i];
             }
           }